Struct swc_ecma_codegen::Config

pub struct Config {
    pub target: EsVersion,
    pub ascii_only: bool,
    pub minify: bool,
    pub omit_last_semi: bool,
}

Fields
target: EsVersion

The target runtime environment.

This defaults to EsVersion::latest because it preserves input as much as possible.

Note: This does not verifies if output is valid for the target runtime. e.g. const foo = 1; with EsVersion::Es3 will emitted as const foo = 1 without verification. This is because it’s not a concern of the code generator.

ascii_only: bool

Forces the code generator to use only ascii characters.

This is useful for environments that do not support unicode.

minify: boolomit_last_semi: bool

If true, the code generator will emit the lastest semicolon.

Defaults to false.