Crate swc_ecma_codegen

Modules

macros

text_writer

util

Structs

Config

Emitter

Traits

Node

Functions

to_code

Generate a code from a syntax node using default options.

to_code_default

Generate a code from a syntax node using default options.

to_code_with_comments

Generate a code from a syntax node using default options.

Type Aliases

Result